Assistant checkpoint: Create MealView page and update navigation flow
Assistant generated file changes:
- client/src/pages/MealView.tsx: Create new MealView page
- client/src/components/meal-plan/MealCell.tsx: Update MealCell to use new navigation
- client/src/components/meal-plan/MealPlanTable.tsx: Update MealPlanTable cell rendering
- client/src/App.tsx: Add MealView route
Add MealView route

---

User prompt:

Create a new page for the Meal, where a user can see the list of selected recipes, can remove recipes or add new ones, make sure the name of the day and the date are present. Remove the modal from the Meal Planner and instead take the user to this new page. If the Meal cell does not have any recipes open the NewRecipe modal, but if it has recipes, take the user to the new Meal page.